Silenced warning by switching to size_t

+++ DispatchProcessingDemo/DispatchFrameworkCore/SimpleSource.cpp
@@ -12,7 +12,7 @@
 
 using namespace demo;
 
-SimpleSource::SimpleSource(unsigned int iNEvents): 
+SimpleSource::SimpleSource(size_t iNEvents): 
   m_nEvents(iNEvents),
   m_seenEvents(0) {}
